+++ packages/liveblocks-zustand/src/index.test.ts
@@ -1,4 +1,4 @@
-import type { JsonObject, Presence } from "@liveblocks/client";
+import type { JsonObject } from "@liveblocks/client";
 import { createClient } from "@liveblocks/client";
 import type {
   IdTuple,
@@ -98,7 +98,7 @@ const basicStateCreator: StateCreator<BasicStore> = (set) => ({
 
 function prepareClientAndStore<
   T extends ZustandState,
-  TPresence extends Presence = Presence
+  TPresence extends JsonObject = JsonObject
 >(
   stateCreator: StateCreator<T>,
   options: {